# Implement unary functionality using streams