--- a/call.go
+++ b/call.go
@@ -43,7 +43,7 @@ func Invoke(ctx context.Context, method string, args, reply interface{}, cc *Cli
 
 var unaryStreamDesc = &StreamDesc{ServerStreams: false, ClientStreams: false}
 
-func invoke(ctx context.Context, method string, req, reply interface{}, cc *ClientConn, opts ...CallOption) (e error) {
+func invoke(ctx context.Context, method string, req, reply interface{}, cc *ClientConn, opts ...CallOption) error {
 	// TODO: implement retries in clientStream and make this simply
 	// newClientStream, SendMsg, RecvMsg.
 	firstAttempt := true