@dfawley 2018-01-25
I ran 5 runs of the unary benchmark with 1 byte request/response and 100 concurrent streams for 2 minutes each.

Before:

```
50_Latency: 1465.2040 µs 	90_Latency: 1635.4170 µs 	99_Latency: 2209.4680 µs 	Avg latency: 1492.8950 µs 	Count: 8034655 	7352 Bytes/op	141 Allocs/op	
50_Latency: 1468.3980 µs 	90_Latency: 1634.9030 µs 	99_Latency: 2206.0530 µs 	Avg latency: 1496.6130 µs 	Count: 8015025 	7352 Bytes/op	141 Allocs/op	
50_Latency: 1460.1300 µs 	90_Latency: 1624.9120 µs 	99_Latency: 2201.9330 µs 	Avg latency: 1487.6050 µs 	Count: 8063496 	7352 Bytes/op	141 Allocs/op	
50_Latency: 1469.1540 µs 	90_Latency: 1633.3620 µs 	99_Latency: 2201.2620 µs 	Avg latency: 1496.0120 µs 	Count: 8018548 	7352 Bytes/op	142 Allocs/op	
50_Latency: 1469.0910 µs 	90_Latency: 1630.6830 µs 	99_Latency: 2203.9370 µs 	Avg latency: 1496.4050 µs 	Count: 8016386 	7352 Bytes/op	142 Allocs/op	
```

Average of averages: 1493

After:
```
50_Latency: 1497.4120 µs 	90_Latency: 1664.6800 µs 	99_Latency: 2243.7390 µs 	Avg latency: 1527.6940 µs 	Count: 7852325 	7720 Bytes/op	145 Allocs/op	
50_Latency: 1498.5790 µs 	90_Latency: 1664.7820 µs 	99_Latency: 2229.4360 µs 	Avg latency: 1526.6400 µs 	Count: 7857824 	7721 Bytes/op	145 Allocs/op	
50_Latency: 1500.0920 µs 	90_Latency: 1667.4590 µs 	99_Latency: 2235.4930 µs 	Avg latency: 1528.6580 µs 	Count: 7846750 	7721 Bytes/op	145 Allocs/op	
50_Latency: 1499.1370 µs 	90_Latency: 1664.3580 µs 	99_Latency: 2229.4210 µs 	Avg latency: 1527.3680 µs 	Count: 7853837 	7720 Bytes/op	145 Allocs/op	
50_Latency: 1490.3740 µs 	90_Latency: 1657.7140 µs 	99_Latency: 2233.4980 µs 	Avg latency: 1519.9220 µs 	Count: 7892692 	7720 Bytes/op	145 Allocs/op	
```

Average of averages: 1508

So with this change we are approximately 1% slower on this unary microbenchmark.  I'll try to see if there's any obvious low-hanging fruit, but my initial thinking is that the 1% is worth it to unify the code (e.g. we can focus our performance work even better with less code to optimize), and it should have very little impact in real-world scenarios.