@menghanl 2018-02-02
Please revert the benchmark commits.

---

Reviewed 1 of 3 files at r1, 2 of 2 files at r3.
Review status: 3 of 7 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[call.go, line 46 at r3](https://reviewable.io:443/reviews/grpc/grpc-go/1835#-L4NPMmTd9vlmHZ5mSTO:-L4NPMmTd9vlmHZ5mSTP:bq93kc9) ([raw file](https://github.com/grpc/grpc-go/blob/f39e079d33fddff68853ed4da535e9bd912afab5/call.go#L46)):*
> ```Go
> var unaryStreamDesc = &StreamDesc{ServerStreams: false, ClientStreams: false}
> 
> func invoke(ctx context.Context, method string, req, reply interface{}, cc *ClientConn, opts ...CallOption) (e error) {
> ```

Returned error doesn't need to be named

---

*[stream.go, line 493 at r3](https://reviewable.io:443/reviews/grpc/grpc-go/1835#-L4NWAYs4x8wDt558lMx:-L4NWAYs4x8wDt558lMy:b9yasm9) ([raw file](https://github.com/grpc/grpc-go/blob/f39e079d33fddff68853ed4da535e9bd912afab5/stream.go#L493)):*
> ```Go
> }
> 
> func (cs *clientStream) CloseSend() (err error) {
> ```

Also, no need to name the returned error

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/grpc/grpc-go/1835#-:-L4NXA7LMRpeyYdXnQFK:becrgv)*
<!-- Sent from Reviewable.io -->