@dfawley 2018-02-05
Benchmark commit reverted via rebase. Hopefully that doesn't break reviewable's/github's brain too much.  And since I was rebasing anyway, I also updated with master at the same time because there was a conflict.

---

Review status: 3 of 7 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[call.go, line 46 at r3](https://reviewable.io:443/reviews/grpc/grpc-go/1835#-L4NPMmTd9vlmHZ5mSTO:-L4azlAv9TOS4EJCzy8N:b-896fix) ([raw file](https://github.com/grpc/grpc-go/blob/f39e079d33fddff68853ed4da535e9bd912afab5/call.go#L46)):*
<details><summary><i>Previously, menghanl (Menghan Li) wrote…</i></summary><blockquote>

Returned error doesn't need to be named
</blockquote></details>

Done.

---

*[stream.go, line 493 at r3](https://reviewable.io:443/reviews/grpc/grpc-go/1835#-L4NWAYs4x8wDt558lMx:-L4azuEnlD6takSiHvyr:b-896fix) ([raw file](https://github.com/grpc/grpc-go/blob/f39e079d33fddff68853ed4da535e9bd912afab5/stream.go#L493)):*
<details><summary><i>Previously, menghanl (Menghan Li) wrote…</i></summary><blockquote>

Also, no need to name the returned error
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/grpc/grpc-go/1835#-:-L4b-AVOmvvfbiZivJrV:bln2986)*
<!-- Sent from Reviewable.io -->